fixed a A bug when when admin processes all KYC requests in the queue, there are
